refactor: migrate xp and level to long

+++ src/main/java/flustix/fluxifyed/database/api/v1/components/xp/GuildLeaderboard.java
@@ -47,7 +47,7 @@ public class GuildLeaderboard {
         public final String username;
         public final String discriminator;
         public final String avatar;
-        public final int xp;
+        public final long xp;
 
         public LeaderboardUserEntry(User user, XPUser xpUser) {
             username = user.getName();
