+++ src/main/java/flustix/fluxifyed/utils/xp/XPUtils.java
@@ -1,32 +1,32 @@
 package flustix.fluxifyed.utils.xp;
 
 public class XPUtils {
-    public static int calculateLevel(int xp) {
+    public static long calculateLevel(long xp) {
         return calculateLevel(xp, "default");
     }
 
     /**
      * Calculates level from xp with different formulas from other bots (for easier adapting)
      */
-    public static int calculateLevel(int xp, String formula) {
+    public static long calculateLevel(long xp, String formula) {
         switch (formula) {
             case "amari" -> {
-                return (int) Math.sqrt((xp - 35) / 20f) + 1;
+                return (long) Math.sqrt((xp - 35) / 20f) + 1;
             }
             default -> {
-                return (int) Math.floor(Math.sqrt(xp / 100f));
+                return (long) Math.floor(Math.sqrt(xp / 100f));
             }
         }
     }
 
-    public static int calculateXP(int level) {
+    public static long calculateXP(long level) {
         return calculateXP(level, "default");
     }
 
     /**
      * Calculates xp from level with different formulas from other bots (for easier adapting)
      */
-    public static int calculateXP(int level, String formula) {
+    public static long calculateXP(long level, String formula) {
         switch (formula) {
             case "amari" -> {
                 return (int)(20 * Math.pow(level - 1, 2) + 35);
